Change include paths to be relative to include/ directory
Previously, all include statements were just the names of the headers,
even for headers in subdirectories, which meant that all
subdirectories needed to be added to the include path. A better way to
do it is to make all include paths relative to the include directory
so the only directory that needs to be added to the include path is
that include directory.